+++ app/Services/WebSocketHandler.bak.ts
@@ -2,10 +2,10 @@ import "websocket-polyfill"
 import { v4 as uuidv4 } from "uuid";
 import Env from "@ioc:Adonis/Core/Env"
 import { WebSocket } from "ws";
-import NostrPool from "./NostrPool";
 import { Filter, Event } from "nostr-tools";
 import WebSocketInstance from "./WebSocketInstance"
-import { EventEmitter } from "node:events";
+import { SimplePool } from "./Simplepool";
+import { Sub } from "../Models/Relay";
 
 interface CustomWebsocket extends WebSocket {
     connectionId: string
@@ -16,14 +16,14 @@ class WebSocketServer {
     public booted: boolean
 
     private _relays: string[]
-    private _pool: typeof NostrPool
+    private _pool: SimplePool
     private _cache: Map<string, string>
-    private _subs: Map<string, EventEmitter>
+    private _subs: Map<string, Sub>
 
     constructor() {
         this.booted = false
         this._relays = [...Env.get('RELAYS').split(',')]
-        this._pool = NostrPool
+        this._pool = new SimplePool()
 
         this._cache = new Map()
         this._subs = new Map()
@@ -32,9 +32,21 @@ class WebSocketServer {
     public async boot() {
         if (this.booted) return
 
-        await this._pool.init()
+        await this.initRelays()
         await this.initWsHandler()
         this.booted = true
+    }
+
+    public async initRelays() {
+        // Initializing relays
+        for (const url of this._relays) {
+            try {
+                const relay = await this._pool.ensureRelay(url)
+                await relay.connect()
+            } catch (_) {
+                // console.error(`Error while initializing relay ${url}`)
+            }
+        }
     }
 
     private async initWsHandler() {
@@ -54,16 +66,16 @@ class WebSocketServer {
                         // Close old subscription if subscriptionId already exists
                         const oldRandomSubscriptionId = this._cache.get(`${socket.connectionId}:${subscriptionId}`)
                         if (oldRandomSubscriptionId) {
-                            this._subs.get(oldRandomSubscriptionId)?.emit('unsubscribe')
+                            this._subs.get(oldRandomSubscriptionId)?.unsub()
                             this._subs.delete(oldRandomSubscriptionId)
                         }
 
                         this._cache.set(`${socket.connectionId}:${subscriptionId}`, randomSubscriptionId)
 
-                        this._subs.set(randomSubscriptionId, await this._pool.sub(
+                        this._subs.set(randomSubscriptionId, this._pool.sub(
                             this._relays,
                             [filters],
-                            randomSubscriptionId
+                            { id: randomSubscriptionId }
                         ))
 
                         this._subs.get(randomSubscriptionId)?.on('event', (data: Event) => {
@@ -78,7 +90,7 @@ class WebSocketServer {
 
                         const randomSubscriptionId = this._cache.get(`${socket.connectionId}:${subscriptionId}`)
                         if (randomSubscriptionId) {
-                            this._subs.get(randomSubscriptionId)?.emit('unsubscribe')
+                            this._subs.get(randomSubscriptionId)?.unsub()
                             this._subs.delete(randomSubscriptionId)
 
                             this._cache.delete(`${socket.connectionId}:${subscriptionId}`)
@@ -86,13 +98,12 @@ class WebSocketServer {
                     }
                     else if (parsed[0] === 'EVENT') {
                         const event = parsed[1] as unknown as Event
-                        const pub = await this._pool.publish(this._relays, event)
-                        pub.on('ok', () => {
-                            socket.send(JSON.stringify(["OK", event.id, true, ""]))
-                        })
-                        pub.on('failed', (reason: string) => {
-                            socket.send(JSON.stringify(["NOTICE", reason]))
-                        })
+                        let pubs = this._pool.publish(this._relays, event)
+                        pubs.forEach(pub =>
+                            pub.on('ok', () => {
+                                socket.send(JSON.stringify(["OK", event.id, true, ""]))
+                            })
+                        )
                     }
                     else {
                         throw new Error(`Invalid event ${data.toString()}`)
@@ -103,7 +114,7 @@ class WebSocketServer {
                         if (key.startsWith(socket.connectionId)) {
                             const randomSubscriptionId = this._cache.get(key)
                             if (randomSubscriptionId) {
-                                this._subs.get(randomSubscriptionId)?.emit('unsubscribe')
+                                this._subs.get(randomSubscriptionId)?.unsub()
                                 this._subs.delete(randomSubscriptionId)
                             }
                             this._cache.delete(key)
@@ -118,7 +129,7 @@ class WebSocketServer {
                     if (key.startsWith(socket.connectionId)) {
                         const randomSubscriptionId = this._cache.get(key)
                         if (randomSubscriptionId) {
-                            this._subs.get(randomSubscriptionId)?.emit('unsubscribe')
+                            this._subs.get(randomSubscriptionId)?.unsub()
                             this._subs.delete(randomSubscriptionId)
                         }
                         this._cache.delete(key)
@@ -128,8 +139,17 @@ class WebSocketServer {
         })
     }
 
-    public getStats() {
-        const relays = NostrPool.getRelaysStatus()
+    public async getRelays(): Promise<{ url: string, connected: boolean }[]> {
+        let relays: { url: string, connected: boolean }[] = []
+        for (const relayUrl of this._relays) {
+            const relay = await this._pool.ensureRelay(relayUrl)
+            relays = [...relays, { url: relay.url, connected: relay.status === 1 ? true : false }]
+        }
+        return relays
+    }
+
+    public async getStats() {
+        const relays = await this.getRelays()
         return {
             connectedClients: WebSocketInstance.ws.clients.size,
             internalInfos: {
