+++ providers/AppProvider.ts
@@ -15,7 +15,7 @@ export default class AppProvider {
     // App is ready
     if (this.app.environment === 'web') {
       await import('../start/socket')
-      // await import('../start/cron')
+      await import('../start/cron')
     }
   }
